% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHierarchFDR.R
\name{addHierarchFDR}
\alias{addHierarchFDR}
\title{FUNCTION_TITLE}
\usage{
addHierarchFDR(pvalues, categs, fdr2control = 0.05,
  fdrmethod_level1 = "BH", fdrmethod_level2 = "BH",
  correctionLevel1 = "BB", quiet = FALSE)
}
\arguments{
\item{pvalues}{PARAM_DESCRIPTION}

\item{categs}{PARAM_DESCRIPTION}

\item{fdr2control}{PARAM_DESCRIPTION, Default: 0.05}

\item{fdrmethod_level1}{PARAM_DESCRIPTION, Default: 'BH'}

\item{fdrmethod_level2}{PARAM_DESCRIPTION, Default: 'BH'}

\item{correctionLevel1}{PARAM_DESCRIPTION, Default: 'BB'}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
