% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKorbinianFDR.R
\name{addKorbinianFDR}
\alias{addKorbinianFDR}
\title{FUNCTION_TITLE}
\usage{
addKorbinianFDR(pvals, mystatistic = "pvalue", showplot = F, ...)
}
\arguments{
\item{pvals}{PARAM_DESCRIPTION}

\item{mystatistic}{PARAM_DESCRIPTION, Default: 'pvalue'}

\item{showplot}{PARAM_DESCRIPTION, Default: F}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
