% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_hk.R
\name{barplot_hk}
\alias{barplot_hk}
\title{FUNCTION_TITLE}
\usage{
barplot_hk(
  x,
  shift = "calc",
  srt = 0,
  dotable = T,
  cext = 1,
  textrotate = 0,
  myround = 3,
  showproz = T,
  colplot = "darkred",
  ...
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{shift}{PARAM_DESCRIPTION, Default: 'calc'}

\item{srt}{PARAM_DESCRIPTION, Default: 0}

\item{dotable}{PARAM_DESCRIPTION, Default: T}

\item{cext}{PARAM_DESCRIPTION, Default: 1}

\item{textrotate}{PARAM_DESCRIPTION, Default: 0}

\item{myround}{PARAM_DESCRIPTION, Default: 3}

\item{showproz}{PARAM_DESCRIPTION, Default: T}

\item{colplot}{PARAM_DESCRIPTION, Default: 'darkred'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
