% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bauePlinkCall.R
\name{bauePlinkCall}
\alias{bauePlinkCall}
\title{FUNCTION_TITLE}
\usage{
bauePlinkCall(on_server = as.character(Sys.info()["sysname"]) == "Linux",
  showMessage1 = T, showMessage2 = T)
}
\arguments{
\item{on_server}{PARAM_DESCRIPTION, Default: as.character(Sys.info()["sysname"]) == "Linux"}

\item{showMessage1}{PARAM_DESCRIPTION, Default: T}

\item{showMessage2}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
