% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci2p.R
\name{ci2p}
\alias{ci2p}
\title{FUNCTION_TITLE}
\usage{
ci2p(Est, u, l)
}
\arguments{
\item{Est}{PARAM_DESCRIPTION}

\item{u}{PARAM_DESCRIPTION}

\item{l}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
