% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverFread.R
\name{cleverFread}
\alias{cleverFread}
\title{FUNCTION_TITLE}
\usage{
cleverFread(name, path, myenvir = .GlobalEnv, ...)
}
\arguments{
\item{name}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION}

\item{myenvir}{PARAM_DESCRIPTION, Default: .GlobalEnv}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
