% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverLoadinGlobalEnv.R
\name{cleverLoadinGlobalEnv}
\alias{cleverLoadinGlobalEnv}
\title{FUNCTION_TITLE}
\usage{
cleverLoadinGlobalEnv(name, path, myenvir = .GlobalEnv,
  number_or_name = NULL, ...)
}
\arguments{
\item{name}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION}

\item{myenvir}{PARAM_DESCRIPTION, Default: .GlobalEnv}

\item{number_or_name}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
