% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colplotLikeExcel.R
\name{colplotLikeExcel}
\alias{colplotLikeExcel}
\title{FUNCTION_TITLE}
\usage{
colplotLikeExcel(plotdat, mycolors = c("dodgerblue2", "white", "red"),
  lowest_colorval = "minimum", middle_colorval = "median",
  highest_colorval = "maximum", xlabel = "", ylabel = "",
  x_axis_pos = "top", myround = 0, userdefined_labels = NULL,
  row_names = NULL)
}
\arguments{
\item{plotdat}{PARAM_DESCRIPTION}

\item{mycolors}{PARAM_DESCRIPTION, Default: c("dodgerblue2", "white", "red")}

\item{lowest_colorval}{PARAM_DESCRIPTION, Default: 'minimum'}

\item{middle_colorval}{PARAM_DESCRIPTION, Default: 'median'}

\item{highest_colorval}{PARAM_DESCRIPTION, Default: 'maximum'}

\item{xlabel}{PARAM_DESCRIPTION, Default: ''}

\item{ylabel}{PARAM_DESCRIPTION, Default: ''}

\item{x_axis_pos}{PARAM_DESCRIPTION, Default: 'top'}

\item{myround}{PARAM_DESCRIPTION, Default: 0}

\item{userdefined_labels}{PARAM_DESCRIPTION, Default: NULL}

\item{row_names}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
