% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAllPackages.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{FUNCTION_TITLE}
\usage{
detachAllPackages()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
