% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countLeadTrailSpaces.R
\name{doBasicCheck}
\alias{doBasicCheck}
\title{FUNCTION_TITLE}
\usage{
doBasicCheck(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
