% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_html.R
\name{dt_html}
\alias{dt_html}
\title{FUNCTION_TITLE}
\usage{
dt_html(df2, zeileninitial = 20, maxstringlength = 25)
}
\arguments{
\item{df2}{PARAM_DESCRIPTION}

\item{zeileninitial}{PARAM_DESCRIPTION, Default: 20}

\item{maxstringlength}{PARAM_DESCRIPTION, Default: 25}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
