% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipseVenn.R
\name{ellipseVenn}
\alias{ellipseVenn}
\title{FUNCTION_TITLE}
\usage{
ellipseVenn(lines = lines, olDF, title = title, labels = labels,
  sub = mysub, main, lcol = lcol, tcex = 1.3, ...)
}
\arguments{
\item{lines}{PARAM_DESCRIPTION, Default: lines}

\item{olDF}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION, Default: title}

\item{labels}{PARAM_DESCRIPTION, Default: labels}

\item{sub}{PARAM_DESCRIPTION, Default: mysub}

\item{main}{PARAM_DESCRIPTION}

\item{lcol}{PARAM_DESCRIPTION, Default: lcol}

\item{tcex}{PARAM_DESCRIPTION, Default: 1.3}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
