% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastermultipleEntries2multipleRows.R
\name{fastermultipleEntries2multipleRows}
\alias{fastermultipleEntries2multipleRows}
\title{FUNCTION_TITLE}
\usage{
fastermultipleEntries2multipleRows(dfori, idToSplit, separator)
}
\arguments{
\item{dfori}{PARAM_DESCRIPTION}

\item{idToSplit}{PARAM_DESCRIPTION}

\item{separator}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
