% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_matrixEQTL.R
\name{fdr_matrixEQTL}
\alias{fdr_matrixEQTL}
\title{FUNCTION_TITLE}
\usage{
fdr_matrixEQTL(p, N)
}
\arguments{
\item{p}{PARAM_DESCRIPTION}

\item{N}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
