% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalizeSkript.R
\name{finalizeSkript}
\alias{finalizeSkript}
\title{FUNCTION_TITLE}
\usage{
finalizeSkript(
  myfilename = filename,
  saveTheImage = F,
  dostr = F,
  mypathwd = pathwd
)
}
\arguments{
\item{myfilename}{PARAM_DESCRIPTION, Default: filename}

\item{saveTheImage}{PARAM_DESCRIPTION, Default: F}

\item{dostr}{PARAM_DESCRIPTION, Default: F}

\item{mypathwd}{PARAM_DESCRIPTION, Default: pathwd}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
