% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formateTimediff.R
\name{formateTimediff}
\alias{formateTimediff}
\title{FUNCTION_TITLE}
\usage{
formateTimediff(timediff, mydigits = 3)
}
\arguments{
\item{timediff}{PARAM_DESCRIPTION}

\item{mydigits}{PARAM_DESCRIPTION, Default: 3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
