% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwriteGz.R
\name{fwriteGz}
\alias{fwriteGz}
\title{FUNCTION_TITLE}
\usage{
fwriteGz(df, filename, gzip = T, delim = "\\t", overwriteGz = T, ...)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{filename}{PARAM_DESCRIPTION}

\item{gzip}{PARAM_DESCRIPTION, Default: T}

\item{delim}{PARAM_DESCRIPTION, Default: '    '}

\item{overwriteGz}{PARAM_DESCRIPTION, Default: T}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
