% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilenamesInFolder.R
\name{getFilenamesInFolder}
\alias{getFilenamesInFolder}
\title{FUNCTION_TITLE}
\usage{
getFilenamesInFolder(erkennungsstring = "\\\\.txt$",
  on_server = Sys.info()["sysname"] == "Linux", wd = getwd())
}
\arguments{
\item{erkennungsstring}{PARAM_DESCRIPTION, Default: '\.txt$'}

\item{on_server}{PARAM_DESCRIPTION, Default: Sys.info()["sysname"] == "Linux"}

\item{wd}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
