% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHaeufigkeit.R
\name{getHaeufigkeit}
\alias{getHaeufigkeit}
\title{FUNCTION_TITLE}
\usage{
getHaeufigkeit(var)
}
\arguments{
\item{var}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
