% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlinkFromImpute.R
\name{getPlinkFromImpute}
\alias{getPlinkFromImpute}
\title{FUNCTION_TITLE}
\usage{
getPlinkFromImpute(snps, chr, geno_fn, sample_fn, outfile, n_threads = 1,
  createPlinkCommandOnly = F, snps_fn = tempfile(), use_ids = "id2")
}
\arguments{
\item{snps}{PARAM_DESCRIPTION}

\item{chr}{PARAM_DESCRIPTION}

\item{geno_fn}{PARAM_DESCRIPTION}

\item{sample_fn}{PARAM_DESCRIPTION}

\item{outfile}{PARAM_DESCRIPTION}

\item{n_threads}{PARAM_DESCRIPTION, Default: 1}

\item{createPlinkCommandOnly}{PARAM_DESCRIPTION, Default: F}

\item{snps_fn}{PARAM_DESCRIPTION, Default: tempfile()}

\item{use_ids}{PARAM_DESCRIPTION, Default: 'id2'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
