% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvis_tooltip.R
\name{ggvis_tooltip}
\alias{ggvis_tooltip}
\title{FUNCTION_TITLE}
\usage{
ggvis_tooltip(mtc, xcol, ycol, tooltipcols, on_click = T)
}
\arguments{
\item{mtc}{PARAM_DESCRIPTION}

\item{xcol}{PARAM_DESCRIPTION}

\item{ycol}{PARAM_DESCRIPTION}

\item{tooltipcols}{PARAM_DESCRIPTION}

\item{on_click}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
