% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grayplot_na.R
\name{grayplot_na}
\alias{grayplot_na}
\title{FUNCTION_TITLE}
\usage{
grayplot_na(x, y = NULL, type = "p", bgcolor = "gray90",
  v_over_h = FALSE, pch = 21, bg = "lightblue", col = "black",
  force = c("none", "x", "y", "both"), ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION, Default: NULL}

\item{type}{PARAM_DESCRIPTION, Default: 'p'}

\item{bgcolor}{PARAM_DESCRIPTION, Default: 'gray90'}

\item{v_over_h}{PARAM_DESCRIPTION, Default: FALSE}

\item{pch}{PARAM_DESCRIPTION, Default: 21}

\item{bg}{PARAM_DESCRIPTION, Default: 'lightblue'}

\item{col}{PARAM_DESCRIPTION, Default: 'black'}

\item{force}{PARAM_DESCRIPTION, Default: c("none", "x", "y", "both")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
