% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh.R
\name{hh}
\alias{hh}
\title{FUNCTION_TITLE}
\usage{
hh(d, mydims = 5)
}
\arguments{
\item{d}{PARAM_DESCRIPTION}

\item{mydims}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
