% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidegrayplot_na.R
\name{hidegrayplot_na}
\alias{hidegrayplot_na}
\title{FUNCTION_TITLE}
\usage{
hidegrayplot_na(x, y, ..., type = "p", hlines = NULL,
  hlines.col = "white", hlines.lty = 1, hlines.lwd = 1,
  vlines = NULL, vlines.col = "white", vlines.lty = 1,
  vlines.lwd = 1, xat = NULL, yat = NULL, bgcolor = "gray90",
  xaxt = "n", yaxt = "n", col.lab = graphics::par("col.lab"),
  xlim = NULL, ylim = NULL, xlab = NULL, ylab = NULL, xname, yname,
  xaxs = "i", yaxs = "i", pch = 21, bg = "lightblue",
  col = "black", las = 1, mgp.x = c(2.6, 0.5, 0), mgp.y = c(2.6,
  0.5, 0), force = c("none", "x", "y", "both"), v_over_h = FALSE,
  na.width = 0.06, na.gap = 0.01, main = "")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: 'p'}

\item{hlines}{PARAM_DESCRIPTION, Default: NULL}

\item{hlines.col}{PARAM_DESCRIPTION, Default: 'white'}

\item{hlines.lty}{PARAM_DESCRIPTION, Default: 1}

\item{hlines.lwd}{PARAM_DESCRIPTION, Default: 1}

\item{vlines}{PARAM_DESCRIPTION, Default: NULL}

\item{vlines.col}{PARAM_DESCRIPTION, Default: 'white'}

\item{vlines.lty}{PARAM_DESCRIPTION, Default: 1}

\item{vlines.lwd}{PARAM_DESCRIPTION, Default: 1}

\item{xat}{PARAM_DESCRIPTION, Default: NULL}

\item{yat}{PARAM_DESCRIPTION, Default: NULL}

\item{bgcolor}{PARAM_DESCRIPTION, Default: 'gray90'}

\item{xaxt}{PARAM_DESCRIPTION, Default: 'n'}

\item{yaxt}{PARAM_DESCRIPTION, Default: 'n'}

\item{col.lab}{PARAM_DESCRIPTION, Default: par("col.lab")}

\item{xlim}{PARAM_DESCRIPTION, Default: NULL}

\item{ylim}{PARAM_DESCRIPTION, Default: NULL}

\item{xlab}{PARAM_DESCRIPTION, Default: NULL}

\item{ylab}{PARAM_DESCRIPTION, Default: NULL}

\item{xname}{PARAM_DESCRIPTION}

\item{yname}{PARAM_DESCRIPTION}

\item{xaxs}{PARAM_DESCRIPTION, Default: 'i'}

\item{yaxs}{PARAM_DESCRIPTION, Default: 'i'}

\item{pch}{PARAM_DESCRIPTION, Default: 21}

\item{bg}{PARAM_DESCRIPTION, Default: 'lightblue'}

\item{col}{PARAM_DESCRIPTION, Default: 'black'}

\item{las}{PARAM_DESCRIPTION, Default: 1}

\item{mgp.x}{PARAM_DESCRIPTION, Default: c(2.6, 0.5, 0)}

\item{mgp.y}{PARAM_DESCRIPTION, Default: c(2.6, 0.5, 0)}

\item{force}{PARAM_DESCRIPTION, Default: c("none", "x", "y", "both")}

\item{v_over_h}{PARAM_DESCRIPTION, Default: FALSE}

\item{na.width}{PARAM_DESCRIPTION, Default: 0.06}

\item{na.gap}{PARAM_DESCRIPTION, Default: 0.01}

\item{main}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
