% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htl.R
\name{htl}
\alias{htl}
\title{FUNCTION_TITLE}
\usage{
htl(df, laenge = 2)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{laenge}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
