% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huebsch.R
\name{huebsch}
\alias{huebsch}
\title{FUNCTION_TITLE}
\usage{
huebsch(x, stellen = 1)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{stellen}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
