% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertest2.R
\name{hypertest2}
\alias{hypertest2}
\title{FUNCTION_TITLE}
\usage{
hypertest2(besondere, gezogene, hintergrund, more = T, unique = T)
}
\arguments{
\item{besondere}{PARAM_DESCRIPTION}

\item{gezogene}{PARAM_DESCRIPTION}

\item{hintergrund}{PARAM_DESCRIPTION}

\item{more}{PARAM_DESCRIPTION, Default: T}

\item{unique}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
