% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSkript.R
\name{initializeSkript}
\alias{initializeSkript}
\title{FUNCTION_TITLE}
\usage{
initializeSkript(
  datatable_lines = 3,
  datatable_nrow_allshow = 10,
  datatable_colwidth = 40L,
  datatable_cores = 10,
  computer = "amanMRO",
  add_libpath = T,
  lib_location = "/net/ifs1/san_projekte/projekte/genstat/07_programme/rpackages/",
  myfilename = NULL
)
}
\arguments{
\item{datatable_lines}{PARAM_DESCRIPTION, Default: 3}

\item{datatable_nrow_allshow}{PARAM_DESCRIPTION, Default: 10}

\item{datatable_colwidth}{PARAM_DESCRIPTION, Default: 40}

\item{datatable_cores}{PARAM_DESCRIPTION, Default: 10}

\item{computer}{PARAM_DESCRIPTION, Default: 'amanMRO'}

\item{add_libpath}{PARAM_DESCRIPTION, Default: T}

\item{lib_location}{PARAM_DESCRIPTION, Default: '/net/ifs1/san_projekte/projekte/genstat/07_programme/rpackages/'}

\item{myfilename}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
