% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionTest.R
\name{interactionTest}
\alias{interactionTest}
\title{FUNCTION_TITLE}
\usage{
interactionTest(mean1, se1, mean2, se2)
}
\arguments{
\item{mean1}{PARAM_DESCRIPTION}

\item{se1}{PARAM_DESCRIPTION}

\item{mean2}{PARAM_DESCRIPTION}

\item{se2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
