% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{FUNCTION_TITLE}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{tol}{PARAM_DESCRIPTION, Default: .Machine$double.eps^0.5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
