% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitrHelper.R
\name{knitrHelper}
\alias{knitrHelper}
\title{FUNCTION_TITLE}
\usage{
knitrHelper(myfilename = filename, mypathwd = pathwd,
  computer = "amanMRO")
}
\arguments{
\item{myfilename}{PARAM_DESCRIPTION, Default: filename}

\item{mypathwd}{PARAM_DESCRIPTION, Default: pathwd}

\item{computer}{PARAM_DESCRIPTION, Default: 'amanMRO'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
