% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obj.R
\name{load_obj}
\alias{load_obj}
\title{FUNCTION_TITLE}
\usage{
load_obj(f, number_or_name = 1)
}
\arguments{
\item{f}{PARAM_DESCRIPTION}

\item{number_or_name}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
