% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEuler.R
\name{makeEuler}
\alias{makeEuler}
\title{Create a Proportional Euler Diagram}
\usage{
makeEuler(
  mytab1,
  mytab2,
  titletext = "",
  legendtext = NULL,
  plotteIntermediates = F,
  createSimplifiedLabels = F,
  object_colname = "object",
  direction_colname = "direction",
  increasingCutoff = 0
)
}
\arguments{
\item{mytab1}{First table with objects and effect sizes to compare}

\item{mytab2}{Secib table with objects and effect sizes to compare}

\item{titletext}{Plot Title, Default: ''}

\item{legendtext}{Plot legend, default is the names of the input tables, if set must be acharacter vector of length 2, Default: NULL}

\item{plotteIntermediates}{for debug only, plot intermediate venn plots, Default: F}

\item{createSimplifiedLabels}{show only Number of objects with identical direction in label of the intersect , Default: F}

\item{object_colname}{Column name in table mytab1 with the objects to compare, Default: 'object'}

\item{direction_colname}{Column name in table mytab2 with the objects to compare, Default: 'direction'}

\item{increasingCutoff}{cutoff of same size, typical 0 for standard metric effect sizes or 1 for Foldchanges, Default: 0}
}
\value{
Returns a plot and the code to generate the plot
}
\description{
Create a 2-way Proportional Euler Diagram
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
 library(eulerr)
library(toolboxH)

tab1 = data.table(object = letters[1:10], direction = rnorm(10))
tab1
tab2 = data.table(object = letters[5:25], direction = rnorm(21))
tab2

plottext = makeEuler(tab1 , tab2 , titletext = "Yes, we can Make a Title great again")
# savePlotInMultiFormats("exampleplot", 4, 3, plotecode = plottext)
}
