% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{marginal_plot}
\alias{marginal_plot}
\title{FUNCTION_TITLE}
\usage{
marginal_plot(x, y, group = NULL, data = NULL, lm_formula = y ~ x,
  bw = "nrd0", alpha = 1, plot_legend = T, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{lm_formula}{PARAM_DESCRIPTION, Default: y ~ x}

\item{bw}{PARAM_DESCRIPTION, Default: 'nrd0'}

\item{alpha}{PARAM_DESCRIPTION, Default: 1}

\item{plot_legend}{PARAM_DESCRIPTION, Default: T}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
