% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_hk.R
\name{match_hk}
\alias{match_hk}
\title{FUNCTION_TITLE}
\usage{
match_hk(x, y, testunique = T, makeunique = F, importcol = NULL,
  showMessages = T, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{testunique}{PARAM_DESCRIPTION, Default: T}

\item{makeunique}{PARAM_DESCRIPTION, Default: F}

\item{importcol}{PARAM_DESCRIPTION, Default: NULL}

\item{showMessages}{PARAM_DESCRIPTION, Default: T}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
