% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveColFront.R
\name{moveColFront}
\alias{moveColFront}
\title{FUNCTION_TITLE}
\usage{
moveColFront(d = dataframe, colname = "colname")
}
\arguments{
\item{d}{PARAM_DESCRIPTION, Default: dataframe}

\item{colname}{PARAM_DESCRIPTION, Default: 'colname'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
