% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{mytable}
\alias{mytable}
\title{FUNCTION_TITLE}
\usage{
mytable(x, mydigits = 1, doprint = T, do_science_output = F)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{mydigits}{PARAM_DESCRIPTION, Default: 1}

\item{doprint}{PARAM_DESCRIPTION, Default: T}

\item{do_science_output}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
