% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicepairs.R
\name{nicepairs}
\alias{nicepairs}
\title{FUNCTION_TITLE}
\usage{
nicepairs(x, punktcol = grDevices::rgb(0, 0, 0, 0.3), punktform = 4,
  punktgroesse = NULL, cortype = "spearman", smoothness = 0.95, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{punktcol}{PARAM_DESCRIPTION, Default: rgb(0, 0, 0, 0.3)}

\item{punktform}{PARAM_DESCRIPTION, Default: 4}

\item{punktgroesse}{PARAM_DESCRIPTION, Default: NULL}

\item{cortype}{PARAM_DESCRIPTION, Default: 'spearman'}

\item{smoothness}{PARAM_DESCRIPTION, Default: 0.95}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
