% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_plot2.R
\name{norm_plot2}
\alias{norm_plot2}
\title{FUNCTION_TITLE}
\usage{
norm_plot2(vektor, x_lim = c(min(vektor) - min(vektor) * 0.5, max(vektor)
  + max(vektor) * 0.1), titelgroesse = 1, mybreaks = NULL,
  vektorname = NULL, ...)
}
\arguments{
\item{vektor}{PARAM_DESCRIPTION}

\item{x_lim}{PARAM_DESCRIPTION, Default: c(min(vektor) - min(vektor) * 0.5, max(vektor) + max(vektor) * 
0.1)}

\item{titelgroesse}{PARAM_DESCRIPTION, Default: 1}

\item{mybreaks}{PARAM_DESCRIPTION, Default: NULL}

\item{vektorname}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
