% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratioWald.R
\name{oddsratioWald}
\alias{oddsratioWald}
\title{FUNCTION_TITLE}
\usage{
oddsratioWald(n00, n01, n10, n11, alpha = 0.05)
}
\arguments{
\item{n00}{PARAM_DESCRIPTION}

\item{n01}{PARAM_DESCRIPTION}

\item{n10}{PARAM_DESCRIPTION}

\item{n11}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
