% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelisiere.R
\name{parallelisiere}
\alias{parallelisiere}
\title{FUNCTION_TITLE}
\usage{
parallelisiere(proc = 3, on_server = Sys.info()["sysname"] == "Linux",
  mkl_threads = 1)
}
\arguments{
\item{proc}{PARAM_DESCRIPTION, Default: 3}

\item{on_server}{PARAM_DESCRIPTION, Default: Sys.info()["sysname"] == "Linux"}

\item{mkl_threads}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
