% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_from_png.R
\name{pdf_from_png}
\alias{pdf_from_png}
\title{FUNCTION_TITLE}
\usage{
pdf_from_png(code2parseOrPlot, pdf_filename, temp_pngfile = tempfile(),
  resolution = 300, weite = 13, laenge = 13, einheiten = "cm",
  as1file = T, ...)
}
\arguments{
\item{code2parseOrPlot}{PARAM_DESCRIPTION}

\item{pdf_filename}{PARAM_DESCRIPTION}

\item{temp_pngfile}{PARAM_DESCRIPTION, Default: tempfile()}

\item{resolution}{PARAM_DESCRIPTION, Default: 300}

\item{weite}{PARAM_DESCRIPTION, Default: 13}

\item{laenge}{PARAM_DESCRIPTION, Default: 13}

\item{einheiten}{PARAM_DESCRIPTION, Default: 'cm'}

\item{as1file}{PARAM_DESCRIPTION, Default: T}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
