% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFromBetaSE.R
\name{pFromBetaSE}
\alias{pFromBetaSE}
\title{FUNCTION_TITLE}
\usage{
pFromBetaSE(beta, se, two.sided = T, returnXminIf0 = F)
}
\arguments{
\item{beta}{PARAM_DESCRIPTION}

\item{se}{PARAM_DESCRIPTION}

\item{two.sided}{PARAM_DESCRIPTION, Default: T}

\item{returnXminIf0}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
