% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorMatrix.R
\name{plotCorMatrix}
\alias{plotCorMatrix}
\title{FUNCTION_TITLE}
\usage{
plotCorMatrix(
  m_psy,
  relaxissize = 2,
  rellabelsize = 15,
  textangle = 30,
  maxcolor = "darkred",
  mincolor = "white",
  rundestellen = 2
)
}
\arguments{
\item{m_psy}{PARAM_DESCRIPTION}

\item{relaxissize}{PARAM_DESCRIPTION, Default: 2}

\item{rellabelsize}{PARAM_DESCRIPTION, Default: 15}

\item{textangle}{PARAM_DESCRIPTION, Default: 30}

\item{maxcolor}{PARAM_DESCRIPTION, Default: 'darkred'}

\item{mincolor}{PARAM_DESCRIPTION, Default: 'white'}

\item{rundestellen}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
