% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotellipse.R
\name{plotellipse}
\alias{plotellipse}
\title{FUNCTION_TITLE}
\usage{
plotellipse(center = c(1, 1), radius = c(1, 2), rotate = 1,
  segments = 360, xlab = "", ylab = "", ...)
}
\arguments{
\item{center}{PARAM_DESCRIPTION, Default: c(1, 1)}

\item{radius}{PARAM_DESCRIPTION, Default: c(1, 2)}

\item{rotate}{PARAM_DESCRIPTION, Default: 1}

\item{segments}{PARAM_DESCRIPTION, Default: 360}

\item{xlab}{PARAM_DESCRIPTION, Default: ''}

\item{ylab}{PARAM_DESCRIPTION, Default: ''}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
