% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSE.R
\name{plotMeanSE}
\alias{plotMeanSE}
\title{FUNCTION_TITLE}
\usage{
plotMeanSE(categs, werte, plotteSD = F, ...)
}
\arguments{
\item{categs}{PARAM_DESCRIPTION}

\item{werte}{PARAM_DESCRIPTION}

\item{plotteSD}{PARAM_DESCRIPTION, Default: F}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
