% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSankey.R
\name{plotSankey}
\alias{plotSankey}
\title{Plot a sankey plot}
\usage{
plotSankey(
  comparematrix,
  spalte4color,
  gap.width = 0.5,
  colorlines = T,
  alpha = 0.65,
  showblocks = F,
  x_axis_size = 0.1,
  ...
)
}
\arguments{
\item{comparematrix}{data.frame or data.table with plotting data, IMPORTANT Restrict data.frame to colums that should be plotted, NAs currently NOT SUPPORTED}

\item{spalte4color}{column name used for coloring the sankey}

\item{gap.width}{PARAM_DESCRIPTION, Default: 0.5}

\item{colorlines}{PARAM_DESCRIPTION, Default: T}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.65}

\item{showblocks}{PARAM_DESCRIPTION, Default: F}

\item{x_axis_size}{PARAM_DESCRIPTION, Default: 0.1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
head(iris)
iris$Petal.Width_categ=cut(iris$Petal.Width, breaks = quantile(iris$Petal.Width,c(0, 0.5,1)), include.lowest=T)
plotSankey(comparematrix = iris[,c('Species','Petal.Width_categ')], spalte4color = 'Species' )
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}},\code{\link[stringr]{str_length}}
 \code{\link[scales]{alpha}}
}
