% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotte3D.R
\name{plotte3D}
\alias{plotte3D}
\title{FUNCTION_TITLE}
\usage{
plotte3D(x, y, z, farbe, mylabels = "", mysize = 0.6)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{z}{PARAM_DESCRIPTION}

\item{farbe}{PARAM_DESCRIPTION}

\item{mylabels}{PARAM_DESCRIPTION, Default: ''}

\item{mysize}{PARAM_DESCRIPTION, Default: 0.6}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
