% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potenzFormate.R
\name{potenzFormate}
\alias{potenzFormate}
\title{FUNCTION_TITLE}
\usage{
potenzFormate(x, showdigits = 1)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{showdigits}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
