% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.ggd.R
\name{qq.ggd}
\alias{qq.ggd}
\title{FUNCTION_TITLE}
\usage{
qq.ggd(pvector, maxx = NULL, maxy = NULL, ...)
}
\arguments{
\item{pvector}{PARAM_DESCRIPTION}

\item{maxx}{PARAM_DESCRIPTION, Default: NULL}

\item{maxy}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
