% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_conf.R
\name{qq_conf}
\alias{qq_conf}
\title{FUNCTION_TITLE}
\usage{
qq_conf(
  x,
  df = 1,
  x.max = "auto",
  main = "QQ plot",
  plotType = "pval",
  sub = paste(""),
  xlab = "Expected",
  ylab = "Observed",
  conc = c(0.025, 0.975),
  overdisp = FALSE,
  trim = 0.5,
  slope.one = T,
  slope.lambda = FALSE,
  thin = c(0.25, 100),
  oor.pch = 24,
  col.shade = "gray",
  ofname = "qqchi.pdf",
  h = 6,
  w = 6,
  printpdf = F,
  myxlim = "auto",
  point_cex = 0.5,
  point_col = 1,
  point_pch = 1,
  ...
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION, Default: 1}

\item{x.max}{PARAM_DESCRIPTION, Default: 'auto'}

\item{main}{PARAM_DESCRIPTION, Default: 'QQ plot'}

\item{plotType}{PARAM_DESCRIPTION, Default: 'pval'}

\item{sub}{PARAM_DESCRIPTION, Default: paste("")}

\item{xlab}{PARAM_DESCRIPTION, Default: 'Expected'}

\item{ylab}{PARAM_DESCRIPTION, Default: 'Observed'}

\item{conc}{PARAM_DESCRIPTION, Default: c(0.025, 0.975)}

\item{overdisp}{PARAM_DESCRIPTION, Default: FALSE}

\item{trim}{PARAM_DESCRIPTION, Default: 0.5}

\item{slope.one}{PARAM_DESCRIPTION, Default: T}

\item{slope.lambda}{PARAM_DESCRIPTION, Default: FALSE}

\item{thin}{PARAM_DESCRIPTION, Default: c(0.25, 100)}

\item{oor.pch}{PARAM_DESCRIPTION, Default: 24}

\item{col.shade}{PARAM_DESCRIPTION, Default: 'gray'}

\item{ofname}{PARAM_DESCRIPTION, Default: 'qqchi.pdf'}

\item{h}{PARAM_DESCRIPTION, Default: 6}

\item{w}{PARAM_DESCRIPTION, Default: 6}

\item{printpdf}{PARAM_DESCRIPTION, Default: F}

\item{myxlim}{PARAM_DESCRIPTION, Default: 'auto'}

\item{point_cex}{PARAM_DESCRIPTION, Default: 0.5}

\item{point_col}{PARAM_DESCRIPTION, Default: 1}

\item{point_pch}{PARAM_DESCRIPTION, Default: 1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
