% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqPlotMinP.R
\name{qqPlotMinP}
\alias{qqPlotMinP}
\title{FUNCTION_TITLE}
\usage{
qqPlotMinP(
  x,
  number_of_tests = 3,
  xlab = expression(Expected ~ ~-log[10](grDevices::italic(p[min]))),
  ylab = expression(Observed ~ ~-log[10](grDevices::italic(p[min]))),
  main = deparse(substitute(x)),
  las = graphics::par("las"),
  subtitle = paste0("(Considering minimum of ", number_of_tests,
    " independent p-value-series)"),
  envelope = 0.95,
  col = grDevices::palette()[1],
  col.lines = grDevices::palette()[2],
  lwd = 2,
  pch = 16,
  cex = graphics::par("cex"),
  labels = if (!is.null(names(x))) names(x) else seq(along = x),
  id.method = "y",
  id.n = if (id.method[1] == "identify") Inf else 0,
  id.cex = 1,
  id.col = grDevices::palette()[1],
  id.location = "lr",
  grid = TRUE,
  ...
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{number_of_tests}{PARAM_DESCRIPTION, Default: 3}

\item{xlab}{PARAM_DESCRIPTION, Default: expression(Expected ~ ~-log[10](italic(p[min])))}

\item{ylab}{PARAM_DESCRIPTION, Default: expression(Observed ~ ~-log[10](italic(p[min])))}

\item{main}{PARAM_DESCRIPTION, Default: deparse(substitute(x))}

\item{las}{PARAM_DESCRIPTION, Default: par("las")}

\item{subtitle}{PARAM_DESCRIPTION, Default: paste0("(Considering minimum of ", number_of_tests, " independent p-value-series)")}

\item{envelope}{PARAM_DESCRIPTION, Default: 0.95}

\item{col}{PARAM_DESCRIPTION, Default: palette()[1]}

\item{col.lines}{PARAM_DESCRIPTION, Default: palette()[2]}

\item{lwd}{PARAM_DESCRIPTION, Default: 2}

\item{pch}{PARAM_DESCRIPTION, Default: 16}

\item{cex}{PARAM_DESCRIPTION, Default: par("cex")}

\item{labels}{PARAM_DESCRIPTION, Default: if (!is.null(names(x))) names(x) else seq(along = x)}

\item{id.method}{PARAM_DESCRIPTION, Default: 'y'}

\item{id.n}{PARAM_DESCRIPTION, Default: if (id.method[1] == "identify") Inf else 0}

\item{id.cex}{PARAM_DESCRIPTION, Default: 1}

\item{id.col}{PARAM_DESCRIPTION, Default: palette()[1]}

\item{id.location}{PARAM_DESCRIPTION, Default: 'lr'}

\item{grid}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
