% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel2.R
\name{read_excel2}
\alias{read_excel2}
\title{FUNCTION_TITLE}
\usage{
read_excel2(fn, sheet = 1, skip = 0, na = "", ...)
}
\arguments{
\item{fn}{PARAM_DESCRIPTION}

\item{sheet}{PARAM_DESCRIPTION, Default: 1}

\item{skip}{PARAM_DESCRIPTION, Default: 0}

\item{na}{PARAM_DESCRIPTION, Default: ''}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
