% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPlinkBim.R
\name{readPlinkBim}
\alias{readPlinkBim}
\title{FUNCTION_TITLE}
\usage{
readPlinkBim(x, useData.table = T, clever = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{useData.table}{PARAM_DESCRIPTION, Default: T}

\item{clever}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
