% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPlinkFam.R
\name{readPlinkFam}
\alias{readPlinkFam}
\title{FUNCTION_TITLE}
\usage{
readPlinkFam(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
