% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSNPTESTsamplefileASdata_table.R
\name{readSNPTESTsamplefileASdata.table}
\alias{readSNPTESTsamplefileASdata.table}
\title{FUNCTION_TITLE}
\usage{
readSNPTESTsamplefileASdata.table(snptest_sample_fn,
  makeBinaryOutcomeNumeric12 = T)
}
\arguments{
\item{snptest_sample_fn}{PARAM_DESCRIPTION}

\item{makeBinaryOutcomeNumeric12}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
