% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyCols.R
\name{removeEmptyCols}
\alias{removeEmptyCols}
\title{FUNCTION_TITLE}
\usage{
removeEmptyCols(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
