% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{FUNCTION_TITLE}
\usage{
round2(x, n)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
