% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlotInMultiFormats.R
\name{savePlotInMultiFormats}
\alias{savePlotInMultiFormats}
\title{Save Plot in multiple Formats}
\usage{
savePlotInMultiFormats(
  corefilename,
  plotecode,
  weite = 7,
  hoehe = 7,
  mymfrow = c(1, 1),
  saveSVG = T,
  savePDF = T,
  saveEPS = T,
  saveJPG = T,
  jpg_res = 300
)
}
\arguments{
\item{corefilename}{filename (optional incl. path) without .*** file extension}

\item{plotecode}{Any code used for plotting a graphic, as String}

\item{weite}{width in inches, Default: 7}

\item{hoehe}{hight in inches, Default: 7}

\item{mymfrow}{mfrow parameter, relevant for base graphic only, Default: c(1, 1)}

\item{saveSVG}{save as SVG, Default: T}

\item{savePDF}{save as PDF, Default: T}

\item{saveEPS}{save as EPS, Default: T}

\item{saveJPG}{save as JPEG, Default: T}

\item{jpg_res}{REsoluation of Jpec, Default: 300}
}
\value{
Files in the specified location
}
\description{
Saves Plot as jpg, pdf, svg, eps the latter two for Adobe Illustrator or freeware Inkscape
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 data(cars)
 savePlotInMultiFormats("Carsplot", "plot(cars$speed ~ cars$dist)")

 #EXAMPLE1
 data(cars)
 require(ggplot2)
 p1 = ggplot(cars, aes(dist, speed)) + geom_point()
 savePlotInMultiFormats("Carsplot", 'plot(p1)')

 }
}
}
