% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_tooltip.R
\name{scatterplot_tooltip}
\alias{scatterplot_tooltip}
\title{FUNCTION_TITLE}
\usage{
scatterplot_tooltip(xx, yy, tooltip, mycolor = "", mysize = 5,
  tooltipprefix = "", ...)
}
\arguments{
\item{xx}{PARAM_DESCRIPTION}

\item{yy}{PARAM_DESCRIPTION}

\item{tooltip}{PARAM_DESCRIPTION}

\item{mycolor}{PARAM_DESCRIPTION, Default: ''}

\item{mysize}{PARAM_DESCRIPTION, Default: 5}

\item{tooltipprefix}{PARAM_DESCRIPTION, Default: ''}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
