% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{shade}
\alias{shade}
\title{FUNCTION_TITLE}
\usage{
shade(x1, y1, x2, y2, color = col.shade)
}
\arguments{
\item{x1}{PARAM_DESCRIPTION}

\item{y1}{PARAM_DESCRIPTION}

\item{x2}{PARAM_DESCRIPTION}

\item{y2}{PARAM_DESCRIPTION}

\item{color}{PARAM_DESCRIPTION, Default: col.shade}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
