% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showStartcode.R
\name{showStartcode}
\alias{showStartcode}
\title{FUNCTION_TITLE}
\usage{
showStartcode(os = "unix", pathwd = getwd(), myfilename = filename)
}
\arguments{
\item{os}{PARAM_DESCRIPTION, Default: 'unix'}

\item{pathwd}{PARAM_DESCRIPTION, Default: getwd()}

\item{myfilename}{PARAM_DESCRIPTION, Default: filename}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
