% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testThePackage.R
\name{testThePackage}
\alias{testThePackage}
\title{FUNCTION_TITLE}
\usage{
testThePackage()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
