% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_peter.R
\name{theme_peter}
\alias{theme_peter}
\title{FUNCTION_TITLE}
\usage{
theme_peter(beschriftungszahlengroesse = 20, relfak = 2)
}
\arguments{
\item{beschriftungszahlengroesse}{PARAM_DESCRIPTION, Default: 20}

\item{relfak}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
