% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn}
\alias{venn}
\title{FUNCTION_TITLE}
\usage{
venn(x, snames = "", ilabels = FALSE, counts = FALSE,
  ellipse = FALSE, zcolor = "style", opacity = 0.3, size = 15,
  cexil = 0.6, cexsn = 0.85, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{snames}{PARAM_DESCRIPTION, Default: ''}

\item{ilabels}{PARAM_DESCRIPTION, Default: FALSE}

\item{counts}{PARAM_DESCRIPTION, Default: FALSE}

\item{ellipse}{PARAM_DESCRIPTION, Default: FALSE}

\item{zcolor}{PARAM_DESCRIPTION, Default: 'style'}

\item{opacity}{PARAM_DESCRIPTION, Default: 0.3}

\item{size}{PARAM_DESCRIPTION, Default: 15}

\item{cexil}{PARAM_DESCRIPTION, Default: 0.6}

\item{cexsn}{PARAM_DESCRIPTION, Default: 0.85}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
