% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn2.R
\name{venn2}
\alias{venn2}
\title{FUNCTION_TITLE}
\usage{
venn2(x1, y1, mytitle = "2-Way Venn Diagram", mylabels = NA,
  plotte = T, venntype = "2")
}
\arguments{
\item{x1}{PARAM_DESCRIPTION}

\item{y1}{PARAM_DESCRIPTION}

\item{mytitle}{PARAM_DESCRIPTION, Default: '2-Way Venn Diagram'}

\item{mylabels}{PARAM_DESCRIPTION, Default: NA}

\item{plotte}{PARAM_DESCRIPTION, Default: T}

\item{venntype}{PARAM_DESCRIPTION, Default: '2'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
