% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn3.R
\name{venn3}
\alias{venn3}
\title{FUNCTION_TITLE}
\usage{
venn3(x1, y1, z1, mytitle = "3-Way Venn Diagram", mylabels = NA,
  plotte = T, venntype = "3")
}
\arguments{
\item{x1}{PARAM_DESCRIPTION}

\item{y1}{PARAM_DESCRIPTION}

\item{z1}{PARAM_DESCRIPTION}

\item{mytitle}{PARAM_DESCRIPTION, Default: '3-Way Venn Diagram'}

\item{mylabels}{PARAM_DESCRIPTION, Default: NA}

\item{plotte}{PARAM_DESCRIPTION, Default: T}

\item{venntype}{PARAM_DESCRIPTION, Default: '3'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
