% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn4.R
\name{venn4}
\alias{venn4}
\title{FUNCTION_TITLE}
\usage{
venn4(x1, y1, z1, w1, mytitle = "4-Way Venn Diagram", mylabels = NA,
  plotte = T, venntype = "4el")
}
\arguments{
\item{x1}{PARAM_DESCRIPTION}

\item{y1}{PARAM_DESCRIPTION}

\item{z1}{PARAM_DESCRIPTION}

\item{w1}{PARAM_DESCRIPTION}

\item{mytitle}{PARAM_DESCRIPTION, Default: '4-Way Venn Diagram'}

\item{mylabels}{PARAM_DESCRIPTION, Default: NA}

\item{plotte}{PARAM_DESCRIPTION, Default: T}

\item{venntype}{PARAM_DESCRIPTION, Default: '4el'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
