% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndiagram.R
\name{venndiagram}
\alias{venndiagram}
\title{FUNCTION_TITLE}
\usage{
venndiagram(x = x, y = y, z = z, w = w, unique = T,
  title = "Venn Diagram", labels = c("x", "y", "z", "w"), lines = 1,
  lcol = 1, tcol = 1, diacol = 1, plot = T, type = "3",
  printsub = TRUE, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: x}

\item{y}{PARAM_DESCRIPTION, Default: y}

\item{z}{PARAM_DESCRIPTION, Default: z}

\item{w}{PARAM_DESCRIPTION, Default: w}

\item{unique}{PARAM_DESCRIPTION, Default: T}

\item{title}{PARAM_DESCRIPTION, Default: 'Venn Diagram'}

\item{labels}{PARAM_DESCRIPTION, Default: c("x", "y", "z", "w")}

\item{lines}{PARAM_DESCRIPTION, Default: 1}

\item{lcol}{PARAM_DESCRIPTION, Default: 1}

\item{tcol}{PARAM_DESCRIPTION, Default: 1}

\item{diacol}{PARAM_DESCRIPTION, Default: 1}

\item{plot}{PARAM_DESCRIPTION, Default: T}

\item{type}{PARAM_DESCRIPTION, Default: '3'}

\item{printsub}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
