% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait4me.R
\name{wait4me}
\alias{wait4me}
\title{FUNCTION_TITLE}
\usage{
wait4me()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
