% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withWarnings.R
\name{withWarnings}
\alias{withWarnings}
\title{FUNCTION_TITLE}
\usage{
withWarnings(fn, sheet = sheet)
}
\arguments{
\item{fn}{PARAM_DESCRIPTION}

\item{sheet}{PARAM_DESCRIPTION, Default: sheet}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
