% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.delim.R
\name{write.delim}
\alias{write.delim}
\title{FUNCTION_TITLE}
\usage{
write.delim(x, y, writeColnames = T, writeRownames = F, createDir = F, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{writeColnames}{PARAM_DESCRIPTION, Default: T}

\item{writeRownames}{PARAM_DESCRIPTION, Default: F}

\item{createDir}{PARAM_DESCRIPTION, Default: F}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
