% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSnptestSamplefile.R
\name{writeSnptestSamplefile}
\alias{writeSnptestSamplefile}
\title{FUNCTION_TITLE}
\usage{
writeSnptestSamplefile(filename, samplefile, vartypes)
}
\arguments{
\item{filename}{PARAM_DESCRIPTION}

\item{samplefile}{PARAM_DESCRIPTION}

\item{vartypes}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
