% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteXLS_hk.R
\name{WriteXLS_hk}
\alias{WriteXLS_hk}
\title{FUNCTION_TITLE}
\usage{
WriteXLS_hk(x, ExcelFileName = "R.xls", SheetNames = x,
  AutoFilter = T, BoldHeaderRow = T, FreezeRow = 1, FreezeCol = 1,
  AdjWidth = F, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{ExcelFileName}{PARAM_DESCRIPTION, Default: 'R.xls'}

\item{SheetNames}{PARAM_DESCRIPTION, Default: x}

\item{AutoFilter}{PARAM_DESCRIPTION, Default: T}

\item{BoldHeaderRow}{PARAM_DESCRIPTION, Default: T}

\item{FreezeRow}{PARAM_DESCRIPTION, Default: 1}

\item{FreezeCol}{PARAM_DESCRIPTION, Default: 1}

\item{AdjWidth}{PARAM_DESCRIPTION, Default: F}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
