% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs_hk.R
\name{xtabs_hk}
\alias{xtabs_hk}
\title{FUNCTION_TITLE}
\usage{
xtabs_hk(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
