% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwriteGz.R
\name{fwriteGz}
\alias{fwriteGz}
\title{Combining fwrite() with gz}
\usage{
fwriteGz(df, filename, gzip = T, delim = "\\t", overwriteGz = T, ...)
}
\arguments{
\item{df}{Dataframe or data.table to save}

\item{filename}{Filename without gz suffix}

\item{gzip}{Whether to gz after writing, Default: T}

\item{delim}{Delimiter, Default is tab}

\item{overwriteGz}{Overwrite an existing gz file Default: T}

\item{...}{passed to fwrite()}
}
\value{
The written and gz compressed file
}
\description{
Wrapper for fwrite and gz from R.utils
}
\details{
Should work on Linux and Windows
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
